flesh out fit function
created a general outline of the steps the function needs to take as well as defining that first step, generating the values for the proposition priors. the two other steps are generating the evidence priors and generating likelihoods. I believe that I will be finished with building the fit function after this and can move on to the prediction function and testing